move dict_mapper tests to raphtory api
