+++ raphtory/src/core/entities/properties/props.rs
@@ -318,84 +318,10 @@ impl PropMapper {
 
 #[cfg(test)]
 mod test {
-    use std::{collections::HashMap, sync::Arc, thread};
-
-    use quickcheck_macros::quickcheck;
-    use rand::seq::SliceRandom;
-    use raphtory_api::core::storage::dict_mapper::DictMapper;
+    use std::{sync::Arc, thread};
     use rayon::prelude::*;
-
     use super::*;
-
-    #[test]
-    fn test_dict_mapper() {
-        let mapper = DictMapper::default();
-        assert_eq!(mapper.get_or_create_id("test"), 0);
-        assert_eq!(mapper.get_or_create_id("test"), 0);
-        assert_eq!(mapper.get_or_create_id("test2"), 1);
-        assert_eq!(mapper.get_or_create_id("test2"), 1);
-        assert_eq!(mapper.get_or_create_id("test"), 0);
-    }
-
-    #[quickcheck]
-    fn check_dict_mapper_concurrent_write(write: Vec<String>) -> bool {
-        let n = 100;
-        let mapper: DictMapper = DictMapper::default();
-
-        // create n maps from strings to ids in parallel
-        let res: Vec<HashMap<String, usize>> = (0..n)
-            .into_par_iter()
-            .map(|_| {
-                let mut ids: HashMap<String, usize> = Default::default();
-                let mut rng = rand::thread_rng();
-                let mut write_s = write.clone();
-                write_s.shuffle(&mut rng);
-                for s in write_s {
-                    let id = mapper.get_or_create_id(s.as_str());
-                    ids.insert(s, id);
-                }
-                ids
-            })
-            .collect();
-
-        // check that all maps are the same and that all strings have been assigned an id
-        let res_0 = &res[0];
-        res[1..n].iter().all(|v| res_0 == v) && write.iter().all(|v| mapper.get_id(v).is_some())
-    }
-
-    // map 5 strings to 5 ids from 4 threads concurrently 1000 times
-    #[test]
-    fn test_dict_mapper_concurrent() {
-        use std::{sync::Arc, thread};
-
-        let mapper = Arc::new(DictMapper::default());
-        let mut threads = Vec::new();
-        for _ in 0..4 {
-            let mapper = Arc::clone(&mapper);
-            threads.push(thread::spawn(move || {
-                for _ in 0..1000 {
-                    mapper.get_or_create_id("test");
-                    mapper.get_or_create_id("test2");
-                    mapper.get_or_create_id("test3");
-                    mapper.get_or_create_id("test4");
-                    mapper.get_or_create_id("test5");
-                }
-            }));
-        }
-
-        for thread in threads {
-            thread.join().unwrap();
-        }
-
-        let mut actual = vec!["test", "test2", "test3", "test4", "test5"]
-            .into_iter()
-            .map(|name| mapper.get_or_create_id(name))
-            .collect::<Vec<_>>();
-        actual.sort();
-
-        assert_eq!(actual, vec![0, 1, 2, 3, 4]);
-    }
-
+    
     #[test]
     fn test_prop_mapper_concurrent() {
         let values = [Prop::I64(1), Prop::U16(0), Prop::Bool(true), Prop::F64(0.0)];
